+++ src/main/kotlin/com/github/rey5137/robotrunnerplugin/actions/RunRobotTestCaseActionGroup.kt
@@ -12,17 +12,24 @@ class RunRobotTestCaseActionGroup : ActionGroup() {
 
     override fun getChildren(e: AnActionEvent?): Array<AnAction> {
         val project = e?.project ?: return emptyArray()
-        val configurationMap =
-            RunManager.getInstance(project).getConfigurationSettingsList(RobotRunConfigurationType::class.java)
-                .map { it.uniqueID to it }
-                .toMap()
+        val configurationMap = RunManager.getInstance(project)
+            .getConfigurationSettingsList(RobotRunConfigurationType::class.java)
+            .associateBy { it.uniqueID }
+
+        val dataContext = e.dataContext
+        val value = PlatformDataKeys.EDITOR.getData(dataContext)?.selectionModel?.selectedText ?: ""
         return RobotRunProjectSettingsState.getInstance(project).settingMap.entries.filter { it.value.testCaseEnable }
             .mapNotNull { configurationMap[it.key] }
-            .map { RunRobotTestCaseAction(it) as AnAction }
+            .map {
+                RunRobotTestCaseAction(
+                    runConfigurationSetting = it,
+                    values = if (value.isEmpty()) emptyList() else listOf(value)
+                ) as AnAction
+            }
             .toMutableList()
             .apply {
                 add(Separator(null))
-                add(CreateRunRobotTestCaseConfigAction())
+                add(CreateRunRobotTestCaseConfigAction(values = if (value.isEmpty()) emptyList() else listOf(value)))
             }
             .toTypedArray()
     }
@@ -33,10 +40,9 @@ class RunRobotTestCaseActionGroup : ActionGroup() {
             || !project.isOpen
             || EditorGutter.KEY.getData(e.dataContext) != null
             || java.lang.Boolean.TRUE == e.dataContext.getData(CommonDataKeys.EDITOR_VIRTUAL_SPACE)
-        ){
+        ) {
             e.presentation.isEnabledAndVisible = false
-        }
-        else {
+        } else {
             val provider = getAvailableCutProvider(e)
             e.presentation.isEnabled = provider != null && provider.isCutEnabled(e.dataContext)
             e.presentation.isVisible = provider != null && provider.isCutVisible(e.dataContext)