+++ src/main/kotlin/com/github/rey5137/robotrunnerplugin/editors/RobotOutputFileEditorProvider.kt
@@ -4,6 +4,7 @@ import com.github.rey5137.robotrunnerplugin.editors.xml.TAG_ROBOT
 import com.intellij.ide.highlighter.XmlFileType
 import com.intellij.ide.scratch.ScratchUtil
 import com.intellij.lang.LanguageUtil
+import com.intellij.lang.xml.XMLLanguage
 import com.intellij.openapi.fileEditor.FileEditorPolicy
 import com.intellij.openapi.fileEditor.WeighedFileEditorProvider
 import com.intellij.openapi.project.Project
@@ -14,7 +15,7 @@ import javax.xml.stream.XMLInputFactory
 class RobotOutputFileEditorProvider : WeighedFileEditorProvider() {
 
     override fun accept(project: Project, file: VirtualFile): Boolean {
-        val isXml = file.fileType == XmlFileType.INSTANCE || (ScratchUtil.isScratch(file) && LanguageUtil.getLanguageForPsi(project, file) == XmlFileType.INSTANCE)
+        val isXml = file.fileType == XmlFileType.INSTANCE || (ScratchUtil.isScratch(file) && LanguageUtil.getLanguageForPsi(project, file) == XMLLanguage.INSTANCE)
         if(!isXml)
             return false