+++ src/main/kotlin/com/github/rey5137/robotrunnerplugin/editors/json/JsonParser.kt
@@ -15,6 +15,7 @@ class JsonParser(val robotElement: RobotElement) {
 
     fun addData(method: String, payload: JSONObject) {
         try {
+//            println("$method - $payload")
             when (method) {
                 METHOD_START_SUITE -> payload.toSuiteElement().apply {
                     currentElement.addSuite(this)
@@ -30,7 +31,41 @@ class JsonParser(val robotElement: RobotElement) {
                     docIndex++,
                     robotElement
                 ).apply {
-                    currentElement.addKeyword(this)
+                    this.tags.addAll(payload.extractStringArray(FIELD_TAGS))
+                    this.arguments.addAll(payload.extractStringArray(FIELD_ARGS))
+                    this.assigns.addAll(payload.extractStringArray(FIELD_ASSIGN))
+
+                    if (this.type == KEYWORD_TYPE_STEP) {
+                        this.updateStepLevel()
+                        val stepKeywords = currentElement.getStepKeywords()
+                        if(stepKeywords != null) {
+                            while(stepKeywords.isNotEmpty() && stepKeywords.last().stepLevel >= this.stepLevel) {
+                                val lastStep = stepKeywords.removeAt(stepKeywords.size - 1)
+                                lastStep.updateStepStatus()
+                            }
+                            if(stepKeywords.isNotEmpty())
+                                stepKeywords.last().addKeyword(this)
+                            stepKeywords.add(this)
+                        }
+                        if(this.parent == null)
+                            currentElement.addKeyword(this)
+                    } else if (!currentElement.isStepKeyword()) {
+                        if (this.type == KEYWORD_TYPE_TEARDOWN) {
+                            currentElement.getStepKeywords()?.let { stepKeywords ->
+                                stepKeywords.asReversed().forEach { it.updateStepStatus() }
+                                stepKeywords.clear()
+                            }
+                            currentElement.addKeyword(this)
+                        } else {
+                            val stepKeywords = currentElement.getStepKeywords()
+                            if (stepKeywords != null && stepKeywords.isNotEmpty())
+                                stepKeywords.last().addKeyword(this)
+                            else
+                                currentElement.addKeyword(this)
+                        }
+                    } else
+                        currentElement.addKeyword(this)
+
                     pushElement(this)
                 }
                 METHOD_LOG_MESSAGE -> payload.toMessageElement(messageIndex++, robotElement).apply {
@@ -43,13 +78,32 @@ class JsonParser(val robotElement: RobotElement) {
                 METHOD_END_TEST -> payload.toStatusElement().apply {
                     (currentElement as TestElement).status = this
                     (currentElement as TestElement).tags.addAll(payload.extractStringArray(FIELD_TAGS))
+                    if(!currentElement.hasTeardownKeywords()) {
+                        currentElement.getStepKeywords()?.let { stepKeywords ->
+                            stepKeywords.asReversed().forEach { it.updateStepStatus() }
+                            stepKeywords.clear()
+                        }
+                    }
                     popElement()
                 }
                 METHOD_END_KEYWORD -> payload.toStatusElement().apply {
-                    (currentElement as KeywordElement).status = this
-                    (currentElement as KeywordElement).tags.addAll(payload.extractStringArray(FIELD_TAGS))
-                    (currentElement as KeywordElement).arguments.addAll(payload.extractStringArray(FIELD_ARGS))
-                    (currentElement as KeywordElement).assigns.addAll(payload.extractStringArray(FIELD_ASSIGN))
+                    val keywordElement = currentElement as KeywordElement
+                    keywordElement.status = this
+                    if(keywordElement.type == KEYWORD_TYPE_STEP) {
+                        keywordElement.status.status = ""
+                        keywordElement.status.endTime = ""
+                    }
+                    else if(keywordElement.type == KEYWORD_TYPE_END_STEP) {
+                        if(keywordElement.parent!!.isStepKeyword()) {
+                            val stepElement = keywordElement.parent as KeywordElement
+                            stepElement.updateStepStatus()
+                            stepElement.findStepRoot()?.getStepKeywords()?.remove(stepElement)
+                        }
+                    }
+                    else if(!keywordElement.hasTeardownKeywords()) {
+                        keywordElement.stepKeywords.asReversed().forEach { it.updateStepStatus() }
+                        keywordElement.stepKeywords.clear()
+                    }
                     popElement()
                 }
             }
@@ -91,6 +145,8 @@ private fun JSONObject.toKeywordElement(
 ): KeywordElement {
     val name = getJSONObject(FIELD_ATTRS).getString(FIELD_KWNAME)
     val library = getJSONObject(FIELD_ATTRS).getString(FIELD_LIBNAME)
+    val isStepKeyword = library == STEP_LIBRARY && name.equals(STEP_KEYWORD, ignoreCase = true)
+    val isEndStepKeyword = library == STEP_LIBRARY && name.equals(END_STEP_KEYWORD, ignoreCase = true)
     val nameIndex = keywordNameMap[name] ?: keywordNameMap.size.apply {
         keywordNameMap[name] = this
         robotElement.keywordNames.add(name)
@@ -99,12 +155,17 @@ private fun JSONObject.toKeywordElement(
         keywordLibMap[library] = this
         robotElement.keywordLibraries.add(library)
     }
+    val type = when {
+        isStepKeyword -> KEYWORD_TYPE_STEP
+        isEndStepKeyword -> KEYWORD_TYPE_END_STEP
+        else -> getJSONObject(FIELD_ATTRS).getString(FIELD_TYPE).toUpperCase()
+    }
     robotElement.docMap[docIndex] = getJSONObject(FIELD_ATTRS).getString(FIELD_DOC)
     return KeywordElement(
         nameIndex = nameIndex,
         libraryIndex = libIndex,
         docIndex = docIndex,
-        type = getJSONObject(FIELD_ATTRS).getString(FIELD_TYPE).toUpperCase(),
+        type = type,
         robotElement = robotElement,
     )
 }
@@ -144,4 +205,12 @@ private fun String.extractMessageTitle(): String {
             return substring(0, it)
     }
     return substring(0, end + 1)
+}
+
+private fun KeywordElement.findStepRoot(): Element? {
+    var element: Element? = this
+    do {
+        element = element?.getParent()
+    } while(element != null && element.isStepKeyword())
+    return element
 }