+++ src/main/kotlin/com/github/rey5137/robotrunnerplugin/editors/ui/TreeNodeWrapper.kt
@@ -2,4 +2,17 @@ package com.github.rey5137.robotrunnerplugin.editors.ui
 
 import javax.swing.tree.DefaultMutableTreeNode
 
-data class TreeNodeWrapper(val node: DefaultMutableTreeNode, val children: MutableList<TreeNodeWrapper>)
+data class TreeNodeWrapper(
+    val node: DefaultMutableTreeNode,
+    val children: MutableList<TreeNodeWrapper>,
+    var stepChildren: MutableList<TreeNodeWrapper>? = null,
+    var parent: TreeNodeWrapper? = null
+) {
+
+    val hasStepChild
+        get() = stepChildren?.isNotEmpty() ?: false
+
+    val lastStepChild
+        get() = stepChildren?.lastOrNull()
+
+}