+++ src/main/kotlin/com/github/rey5137/robotrunnerplugin/editors/ui/argument/InputTableCellRenderer.kt
@@ -19,7 +19,7 @@ class InputTableCellRenderer(private val argumentModel: ArgumentModel) : TableCe
 
     private val stringCellRenderer = object : ColoredTableCellRenderer() {
         override fun customizeCellRenderer(
-            table: JTable?,
+            table: JTable,
             value: Any?,
             selected: Boolean,
             hasFocus: Boolean,