+++ src/main/kotlin/com/github/rey5137/robotrunnerplugin/runconfigurations/actions/OpenOutputFileAction.kt
@@ -7,7 +7,6 @@ import com.github.rey5137.robotrunnerplugin.runconfigurations.text
 import com.intellij.execution.process.ProcessHandler
 import com.intellij.execution.ui.ConsoleView
 import com.intellij.icons.AllIcons
-import com.intellij.openapi.actionSystem.AnAction
 import com.intellij.openapi.actionSystem.AnActionEvent
 import com.intellij.openapi.project.DumbAwareAction
 import com.intellij.openapi.project.Project
@@ -17,8 +16,7 @@ class OpenOutputFileAction(
     private val processHandler: ProcessHandler,
     private val console: ConsoleView
 ) :
-    DumbAwareAction(MyBundle.message("robot.run.configuration.label.open-output"), null, AllIcons.Actions.Menu_open),
-    AnAction.TransparentUpdate {
+    DumbAwareAction(MyBundle.message("robot.run.configuration.label.open-output"), null, AllIcons.Actions.ShowCode) {
     override fun update(e: AnActionEvent) {
         e.presentation.isEnabled = processHandler.isProcessTerminated
     }